Complete get /foods/:id endpoint
Adds necessary function to response such that a request to
/api/foods/:id will return a JSON object containing the
foods 'name' and 'calories in a food object. Tests are
all currently passing. Testing for the endpoint includes
assertions on the happy and sad path status codes as well
as the content of the response body in a successful query.
All tests are currently passing.